+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.5.0 (2018-3-19)
+
 * Add `oao bump` (upgrades a dependency across all sub-packages) (#28).
 
 ## 1.5.0-beta.0 (2018-3-14)